Restructured all the data for lab0, lab1, and completed Task 1
