+++ .idea/encodings.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/OOP/Task1/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/OOP/Task1/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/OOP/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/OOP/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src1/Lab1/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src1/Lab1/resources" charset="UTF-8" />
+  </component>
+</project>
