--- lab-papers-please/classification.md
@@ -1,93 +1,0 @@
-# Classification values
-
-**isHuman**: 
-- true, 
-- false 
-
-**planets**: 
-- Earth, 
-- Asgard, 
-- Betelgeuse, 
-- Vogsphere, 
-- Kashyyyk, 
-- Endor
-
-**age**: characterized by ranges
-
-**traits:**
-- HAIRY,
-- TALL,
-- SHORT,
-- BLONDE,
-- EXTRA_ARMS,
-- EXTRA_HEAD,
-- GREEN,
-- BULKY,
-- POINTY_EARS,
-
-# Classification by Universe
-
-## Star Wars Universe
-(Wookie)
-
-```
-isHumanoid: false
-planet: Kashyyk
-age: 0-400
-traits: HAIRY, TALL
-```
-
-(Ewok)
-
-```
-isHumanoid: false
-planet: Endor
-age: 0-60
-traits: SHORT, HAIRY
-```
-
-## Marvel Universe
-(Asgardian)
-
-```
-isHumanoid: true
-planet: Asgard
-age: 0-5000
-traits: BLONDE, TALL
-```
-
-## Hitchhiker's Universe
-
-(Betelgeusian)
-```
-isHumanoid: true
-planet: BETELGEUSE
-age: 0-100
-traits: EXTRA_ARMS, EXTRA_HEAD
-```
-
-(Vogons)
-```
-isHumanoid: false
-planet: Vogsphere
-age: 0-200
-traits: GREEN, BULKY
-```
-
-# Lord of the rings
-
-(Elf)
-```
-isHumanoid: true
-planet: Earth
-age: indefinite
-traits: BLONDE, POINTY_EARS
-```
-
-(Dwarf)
-```
-isHumanoid: true
-planet: Earth
-age: 0-200
-traits: SHORT, BULKY
-```
