--- lab-papers-please/kotlin-classification/gradle.properties
@@ -1,1 +1,0 @@
-kotlin.code.style=official
