--- lab-papers-please/output/hitchhiker.json
@@ -1,27 +1,0 @@
-{
-  "name": "hitchHiker",
-  "individuals": [
-    {
-      "id": 2,
-      "isHumanoid": false,
-      "age": 128,
-      "traits": [
-        "GREEN"
-      ]
-    },
-    {
-      "id": 4,
-      "age": 26,
-      "traits": [
-        "EXTRA_ARMS"
-      ]
-    },
-    {
-      "id": 5,
-      "isHumanoid": false,
-      "traits": [
-        "BULKY"
-      ]
-    }
-  ]
-}
