--- lab-papers-please/output/marvel.json
@@ -1,1 +1,0 @@
-{"name":"starwars","individuals":[{"id":1,"isHumanoid":false,"planet":"Endor","age":22}]}
