--- lab-papers-please/output/rings.json
@@ -1,1 +1,0 @@
-{"name":"rings","individuals":[{"id":3,"isHumanoid":true,"age":7523}]}
