--- lab-papers-please/output/starwars.json
@@ -1,11 +1,0 @@
-{
-  "name": "starwars",
-  "individuals": [
-    {
-      "id": 1,
-      "isHumanoid": false,
-      "planet": "Endor",
-      "age": 22
-    }
-  ]
-}
