--- lab-papers-please/rust-classification/.gitignore
@@ -1,3 +1,0 @@
-/target/
-
-
