--- lab-papers-please/rust-classification/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "rust-classification"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-serde = {version = "1.0.210", features= ["derive"]}
-serde_json = "1.0.128"
