--- lab-papers-please/rust-classification/resources/input.json
@@ -1,188 +1,0 @@
-{
-  "data": [
-    {
-      "id": 0,
-      "isHumanoid": false,
-      "planet": "Kashyyyk",
-      "age": 253,
-      "traits": [
-        "HAIRY",
-        "TALL"
-      ]
-    },
-    {
-      "id": 1,
-      "isHumanoid": true,
-      "planet": "Betelgeuse",
-      "age": 59,
-      "traits": [
-        "EXTRA_ARMS",
-        "EXTRA_HEAD"
-      ]
-    },
-    {
-      "id": 2,
-      "isHumanoid": false,
-      "traits": [
-        "BULKY"
-      ]
-    },
-    {
-      "id": 3,
-      "isHumanoid": false,
-      "age": 253,
-      "traits": [
-        "HAIRY",
-        "TALL"
-      ]
-    },
-    {
-      "id": 4,
-      "planet": "Earth",
-      "traits": [
-        "SHORT",
-        "BULKY"
-      ]
-    },
-    {
-      "id": 5,
-      "isHumanoid": true,
-      "traits": [
-        "BULKY"
-      ]
-    },
-    {
-      "id": 6,
-      "age": 26,
-      "traits": [
-        "EXTRA_ARMS"
-      ]
-    },
-    {
-      "id": 7,
-      "isHumanoid": true,
-      "planet": "Asgard",
-      "age": 2034,
-      "traits": [
-        "BLONDE",
-        "TALL"
-      ]
-    },
-    {
-      "id": 8,
-      "age": 82,
-      "traits": [
-        "EXTRA_HEAD"
-      ]
-    },
-    {
-      "id": 9,
-      "isHumanoid": true,
-      "planet": "Earth",
-      "age": 124,
-      "traits": [
-        "SHORT",
-        "BULKY"
-      ]
-    },
-    {
-      "id": 10,
-      "planet": "Endor"
-    },
-    {
-      "id": 11,
-      "isHumanoid": false,
-      "age": 128,
-      "traits": [
-        "GREEN"
-      ]
-    },
-    {
-      "id": 12,
-      "isHumanoid": false,
-      "planet": "Endor",
-      "age": 34,
-      "traits": [
-        "HAIRY",
-        "SHORT"
-      ]
-    },
-    {
-      "id": 13,
-      "planet": "Kashyyyk",
-      "age": 352,
-      "traits": [
-        "HAIRY",
-        "TALL"
-      ]
-    },
-    {
-      "id": 14,
-      "isHumanoid": true,
-      "age": 7523
-    },
-    {
-      "id": 15,
-      "isHumanoid": true,
-      "planet": "Earth",
-      "age": 451,
-      "traits": [
-        "BLONDE",
-        "TALL"
-      ]
-    },
-    {
-      "id": 16,
-      "isHumanoid": false,
-      "planet": "Vogsphere",
-      "age": 128,
-      "traits": [
-        "GREEN",
-        "BULKY"
-      ]
-    },
-    {
-      "id": 17,
-      "isHumanoid": true,
-      "age": 154,
-      "traits": [
-        "BLONDE",
-        "TALL"
-      ]
-    },
-    {
-      "id": 18,
-      "isHumanoid": true,
-      "planet": "Asgard",
-      "traits": [
-        "BLONDE",
-        "TALL"
-      ]
-    },
-    {
-      "id": 19,
-      "isHumanoid": true,
-      "planet": "Earth",
-      "age": 451,
-      "traits": [
-        "BLONDE",
-        "POINTY_EARS"
-      ]
-    },
-    {
-      "id": 20,
-      "isHumanoid": false,
-      "planet": "Endor",
-      "age": 22
-    },
-    {
-      "id": 21,
-      "planet": "Betelgeuse",
-      "age": 45,
-      "traits": [
-        "EXTRA_ARMS",
-        "EXTRA_HEAD"
-      ]
-    }
-  ]
-}
