--- lab-papers-please/rust-classification/resources/output/hitch_hicker.json
@@ -1,1 +1,0 @@
-[{"id":1,"isHumanoid":false,"planet":"Endor","age":22,"traits":null},{"id":5,"isHumanoid":false,"planet":null,"age":null,"traits":["BULKY"]}]
