--- lab-papers-please/rust-classification/resources/output/marvel.json
@@ -1,1 +1,0 @@
-[{"id":3,"isHumanoid":true,"planet":null,"age":7523,"traits":null}]
