--- lab-papers-please/rust-classification/resources/output/rings.json
@@ -1,1 +1,0 @@
-[{"id":2,"isHumanoid":false,"planet":null,"age":128,"traits":["GREEN"]},{"id":4,"isHumanoid":null,"planet":null,"age":26,"traits":["EXTRA_ARMS"]}]
