--- lab-papers-please/rust-classification/resources/output/star-wars.json
@@ -1,1 +1,0 @@
-[{"id":0,"isHumanoid":true,"planet":"Asgard","age":2034,"traits":["BLONDE","TALL"]}]
