+++ src/main/resources/output/hitchhikers.json
@@ -1,0 +1,1 @@
+{"name":"Hitchhiker's Guide to the Galaxy","individuals":["{\"id\":3,\"isHumanoid\":true,\"planet\":\"Betelgeuse\",\"age\":59,\"traits\":[\"EXTRA_ARMS\",\"EXTRA_HEAD\"]}","{\"id\":4,\"isHumanoid\":false,\"planet\":\"Vogsphere\",\"age\":128,\"traits\":[\"GREEN\",\"BULKY\"]}","{\"id\":8,\"isHumanoid\":false,\"planet\":\"Endor\",\"age\":22}","{\"id\":10,\"planet\":\"Betelgeuse\",\"age\":45,\"traits\":[\"EXTRA_ARMS\",\"EXTRA_HEAD\"]}","{\"id\":11,\"isHumanoid\":false,\"age\":128,\"traits\":[\"GREEN\"]}","{\"id\":12,\"isHumanoid\":true,\"age\":7523}","{\"id\":13,\"planet\":\"Earth\",\"traits\":[\"SHORT\",\"BULKY\"]}","{\"id\":15,\"planet\":\"Endor\"}","{\"id\":16,\"isHumanoid\":true,\"age\":154,\"traits\":[\"BLONDE\",\"TALL\"]}","{\"id\":17,\"age\":26,\"traits\":[\"EXTRA_ARMS\"]}","{\"id\":18,\"isHumanoid\":false,\"traits\":[\"BULKY\"]}","{\"id\":20,\"isHumanoid\":true,\"traits\":[\"BULKY\"]}","{\"id\":21,\"age\":82,\"traits\":[\"EXTRA_HEAD\"]}"]}
