+++ src/main/resources/output/marvel.json
@@ -1,0 +1,1 @@
+{"name":"Marvel","individuals":["{\"id\":5,\"isHumanoid\":true,\"planet\":\"Earth\",\"age\":451,\"traits\":[\"BLONDE\",\"POINTY_EARS\"]}","{\"id\":6,\"isHumanoid\":true,\"planet\":\"Earth\",\"age\":124,\"traits\":[\"SHORT\",\"BULKY\"]}","{\"id\":19,\"isHumanoid\":true,\"planet\":\"Earth\",\"age\":451,\"traits\":[\"BLONDE\",\"TALL\"]}"]}
