+++ src/main/resources/output/rings.json
@@ -1,0 +1,1 @@
+{"name":"Rings","individuals":["{\"id\":2,\"isHumanoid\":true,\"planet\":\"Asgard\",\"age\":2034,\"traits\":[\"BLONDE\",\"TALL\"]}","{\"id\":9,\"isHumanoid\":true,\"planet\":\"Asgard\",\"traits\":[\"BLONDE\",\"TALL\"]}"]}
