+++ src/main/resources/output/starwars.json
@@ -1,0 +1,1 @@
+{"name":"Star Wars","individuals":["{\"id\":0,\"isHumanoid\":false,\"planet\":\"Kashyyyk\",\"age\":253,\"traits\":[\"HAIRY\",\"TALL\"]}","{\"id\":1,\"isHumanoid\":false,\"planet\":\"Endor\",\"age\":34,\"traits\":[\"HAIRY\",\"SHORT\"]}","{\"id\":7,\"planet\":\"Kashyyyk\",\"age\":352,\"traits\":[\"HAIRY\",\"TALL\"]}","{\"id\":14,\"isHumanoid\":false,\"age\":253,\"traits\":[\"HAIRY\",\"TALL\"]}"]}
